file with no variables called in command line

+++ ex12.rb
@@ -6,7 +6,7 @@ puts "A bigger number is #{bigger}."
 
 print "Give me another number: "
 another = gets.chomp
-number = another.to_i
+number = another.to_f 
 
 smaller = number / 100
 puts "A smaller number is #{smaller}."
